Add API key auth config
